ltc-insurance-platform/backend/requirements.txt
# LTC Insurance Platform - Backend Requirements

# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Snowflake and Data Processing
snowflake-snowpark-python==1.11.1
pandas==2.1.3
numpy==1.26.2

# Redis Caching
redis==5.0.1

# Configuration and Validation
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
urllib3==2.1.0

# Logging and Monitoring
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type Checking (optional)
mypy==1.7.1
types-redis==4.6.0.11

